refactor(runtime): split driver_proc into handshake/lifecycle/log_forward submodules
Mechanical split of crates/midori-runtime/src/driver_proc.rs (~810 lines)
into a directory module with four files grouped by responsibility:

- mod.rs: SpawnOptions, DriverHandle (incl. Drop), SpawnError,
  spawn_driver / spawn_driver_with_options, public constants
- handshake.rs: HelloMessage / HelloAckMessage / read_hello_line /
  write_hello_ack / is_sdk_compatible / Compatibility /
  HandshakeReadOutcome / json_type_error
- lifecycle.rs: terminate_with_grace / send_sigterm / send_sigkill /
  kill_and_reap / spawn_exit_watcher
- log_forward.rs: LogEvent / non_json_line_to_log_event /
  spawn_log_forwarder

No behavior change. No new dependencies. No public-API changes — every
previously-public item remains public at its existing import path.
Cross-module visibility uses pub(super) to keep the new surface minimal.

Closes #62
