+++ crates/midori-runtime/src/driver_proc/mod.rs
@@ -56,13 +56,26 @@
 //! Tests construct handles via [`spawn_driver_with_options`] with a short
 //! grace period to keep the suite fast.
 
-use std::io::{BufRead, BufReader, Write};
+use std::io::Write;
 use std::path::PathBuf;
-use std::process::{Child, ChildStdin, ChildStdout, Command, Stdio};
+use std::process::{ChildStdin, Command, Stdio};
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{mpsc, Arc};
-use std::thread::{self, JoinHandle};
-use std::time::{Duration, Instant};
+use std::sync::Arc;
+use std::thread::JoinHandle;
+use std::time::Duration;
+
+mod handshake;
+mod lifecycle;
+mod log_forward;
+
+pub use log_forward::{non_json_line_to_log_event, LogEvent};
+
+use handshake::{
+    is_sdk_compatible, json_type_error, read_hello_line, write_hello_ack, Compatibility,
+    HelloAckMessage, HelloMessage,
+};
+use lifecycle::{kill_and_reap, spawn_exit_watcher, terminate_with_grace};
+use log_forward::spawn_log_forwarder;
 
 /// Default upper bound on how long [`spawn_driver`] waits for the driver to
 /// emit its `hello` line. Drivers are expected to write `hello` synchronously
@@ -77,37 +90,6 @@ pub const HANDSHAKE_TIMEOUT: Duration = Duration::from_secs(5);
 /// above the realistic exit budget while still putting an outer bound on
 /// how long a misbehaving driver can stall bridge shutdown.
 pub const DEFAULT_SHUTDOWN_GRACE: Duration = Duration::from_secs(3);
-
-/// SDK majors the Bridge currently understands. The compatibility check
-/// accepts any `hello.sdk_version` parseable as `MAJOR.MINOR.PATCH` whose
-/// `MAJOR` is listed here. Bumping the SDK's binary layout extends this list
-/// (or replaces it) — see `design/10-driver-plugin.md`「バージョン互換性」 for
-/// the policy: layout changes happen on semver-major bumps only.
-const ACCEPTED_SDK_MAJORS: &[u32] = &[0, 1];
-
-/// Poll interval used while waiting for the child to exit during the SIGTERM
-/// → SIGKILL escalation window. Small enough that the worst-case overshoot
-/// of the grace deadline is negligible (50 ms on a 3 s grace), large enough
-/// that the busy-wait does not chew CPU.
-const SHUTDOWN_POLL_INTERVAL: Duration = Duration::from_millis(50);
-
-/// JSON-tagged `hello` message read from the driver's stdout.
-#[derive(Debug, serde::Deserialize)]
-struct HelloMessage {
-    #[serde(rename = "type")]
-    message_type: String,
-    sdk_version: String,
-}
-
-/// JSON-tagged `hello_ack` message written to the driver's stdin.
-#[derive(Debug, serde::Serialize)]
-struct HelloAckMessage<'a> {
-    #[serde(rename = "type")]
-    message_type: &'a str,
-    compatible: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    reason: Option<&'a str>,
-}
 
 /// Tunable options for [`spawn_driver_with_options`]. Held as a struct so
 /// adding more fields (e.g. environment, working directory) does not blow
@@ -462,349 +444,4 @@ pub fn spawn_driver_with_options(
         log_forwarder: Some(log_forwarder),
         watcher: Some(watcher),
     })
-}
-
-/// Internal channel payload used during handshake so the worker thread can
-/// hand its `BufReader<ChildStdout>` back when it returns the first line.
-/// Transferring ownership lets the post-handshake log forwarder pick up
-/// reading from byte 0 of the next line without re-opening stdout.
-enum HandshakeReadOutcome {
-    Line {
-        line: String,
-        reader: BufReader<ChildStdout>,
-    },
-    Err(std::io::Error),
-}
-
-/// Read the driver's first stdout line under a wall-clock timeout, handing
-/// the reader back to the caller so the post-handshake forwarder can keep
-/// reading from byte 0 of line 2.
-///
-/// On the error paths the child is killed and reaped before returning so
-/// the worker thread's `read_line` unblocks (otherwise `join` would deadlock
-/// against the read still waiting on the live pipe).
-fn read_hello_line(
-    stdout: ChildStdout,
-    timeout: Duration,
-    child: &mut Child,
-) -> Result<(String, BufReader<ChildStdout>), SpawnError> {
-    let (tx, rx) = mpsc::channel::<HandshakeReadOutcome>();
-    let reader_handle = thread::spawn(move || {
-        let mut reader = BufReader::new(stdout);
-        let mut line = String::new();
-        match reader.read_line(&mut line) {
-            Ok(0) => {
-                // Sender drops on return; the receiver sees Disconnected.
-            }
-            Ok(_) => {
-                let _ = tx.send(HandshakeReadOutcome::Line { line, reader });
-            }
-            Err(err) => {
-                let _ = tx.send(HandshakeReadOutcome::Err(err));
-            }
-        }
-    });
-
-    let result = match rx.recv_timeout(timeout) {
-        Ok(HandshakeReadOutcome::Line { line, reader }) => Ok((line, reader)),
-        Ok(HandshakeReadOutcome::Err(io_err)) => {
-            kill_and_reap(child);
-            Err(SpawnError::Spawn(io_err))
-        }
-        Err(mpsc::RecvTimeoutError::Timeout) => {
-            kill_and_reap(child);
-            Err(SpawnError::HandshakeTimeout(timeout))
-        }
-        Err(mpsc::RecvTimeoutError::Disconnected) => {
-            kill_and_reap(child);
-            Err(SpawnError::HandshakeMissing)
-        }
-    };
-    let _ = reader_handle.join();
-    result
-}
-
-/// Convert a non-JSON driver stdout line into the structured fields the
-/// bridge logger expects. Pure function: no IO, no globals — the integration
-/// test path verifies the call site is wired, and this unit-tests the
-/// translation without needing a process boundary.
-#[must_use]
-pub fn non_json_line_to_log_event(driver_name: &str, line: &str) -> LogEvent {
-    LogEvent {
-        layer: format!("driver/{driver_name}"),
-        device: None,
-        message: line.trim_end_matches(['\r', '\n']).to_owned(),
-    }
-}
-
-/// Output of [`non_json_line_to_log_event`]. Matches the field set the
-/// `crate::logging::info` entrypoint takes.
-#[derive(Debug, PartialEq, Eq)]
-pub struct LogEvent {
-    pub layer: String,
-    pub device: Option<String>,
-    pub message: String,
-}
-
-/// Drive the post-handshake stdout reader on its own thread, forwarding
-/// each line into the bridge logger. Today every line is treated as a log
-/// line — driver→bridge JSON control messages do not exist yet. When they
-/// do, the dispatch will branch here.
-fn spawn_log_forwarder(driver_name: String, mut reader: BufReader<ChildStdout>) -> JoinHandle<()> {
-    thread::spawn(move || {
-        let mut line = String::new();
-        loop {
-            line.clear();
-            match reader.read_line(&mut line) {
-                Ok(0) => return, // EOF: child closed stdout
-                Ok(_) => {
-                    let event = non_json_line_to_log_event(&driver_name, &line);
-                    crate::logging::info(&event.layer, event.device.as_deref(), &event.message);
-                }
-                Err(_err) => {
-                    // Read errors mean the pipe is gone; the watcher will
-                    // observe the child exit and flip `alive`. Bail.
-                    return;
-                }
-            }
-        }
-    })
-}
-
-/// Wait on the child process from a dedicated thread so the bridge learns
-/// about abnormal exits without blocking foreground work, and so the
-/// [`Drop`]-side SIGTERM / SIGKILL escalation can synchronize against the
-/// child's reap purely through the `alive` flag and the eventual thread
-/// join. The watcher fully owns the `Child` for the duration of its life;
-/// no other code ever calls `wait()` on it.
-fn spawn_exit_watcher(
-    driver_name: String,
-    mut child: Child,
-    alive: Arc<AtomicBool>,
-    shutdown_initiated: Arc<AtomicBool>,
-) -> JoinHandle<()> {
-    thread::spawn(move || {
-        let status = child.wait();
-        alive.store(false, Ordering::SeqCst);
-
-        // If the bridge initiated the shutdown (i.e. `Drop` ran and sent
-        // SIGTERM / SIGKILL), the resulting exit is **expected** even when
-        // it manifests as a non-zero status (signal-terminated processes
-        // report `ExitStatus::success() == false` on Unix). Suppressing
-        // the abnormal-exit log on this path keeps "normal driver
-        // shutdown" out of the bridge error stream.
-        if shutdown_initiated.load(Ordering::SeqCst) {
-            return;
-        }
-
-        match status {
-            Ok(status) if status.success() => {
-                // Graceful exit: nothing to log; the handle owner will
-                // observe `is_alive() == false` if it cares.
-            }
-            Ok(status) => {
-                crate::logging::error(
-                    "bridge",
-                    Some(&driver_name),
-                    format_args!("driver subprocess exited abnormally: {status}"),
-                );
-            }
-            Err(err) => {
-                crate::logging::error(
-                    "bridge",
-                    Some(&driver_name),
-                    format_args!("driver subprocess wait() failed: {err}"),
-                );
-            }
-        }
-    })
-}
-
-/// SIGTERM → grace → SIGKILL escalation. Synchronizes with the watcher
-/// thread purely through the shared `alive` flag: the watcher sets it to
-/// `false` after `child.wait()` returns, so polling it tells us when the
-/// child has been reaped without us having to share the `Child` itself.
-fn terminate_with_grace(alive: &AtomicBool, child_pid: u32, grace: Duration) {
-    // Fast path: child already gone (graceful exit triggered by stdin EOF
-    // closing the SDK loop, or pre-existing crash). Nothing to do.
-    if !alive.load(Ordering::SeqCst) {
-        return;
-    }
-
-    // Step 1: send SIGTERM. Best-effort; ESRCH (child raced to exit
-    // between the load above and the kill below) is ignored inside
-    // `send_sigterm`.
-    send_sigterm(child_pid);
-
-    // Step 2: poll the watcher's `alive` flag. The watcher sets it to
-    // `false` after `child.wait()` returns, so a `false` reading means the
-    // kernel has reaped the child and there is no SIGKILL work to do.
-    let deadline = Instant::now() + grace;
-    while alive.load(Ordering::SeqCst) {
-        if Instant::now() >= deadline {
-            // Step 3: SIGKILL escalation. Sending the signal via the cached
-            // PID is sufficient — the watcher is parked in `child.wait()`
-            // and will return as soon as the kernel delivers the SIGKILL.
-            // The watcher then flips `alive` and exits, and the caller's
-            // `Drop` joins it next, which synchronizes the reap.
-            send_sigkill(child_pid);
-            return;
-        }
-        thread::sleep(SHUTDOWN_POLL_INTERVAL);
-    }
-}
-
-/// Send `SIGTERM` to `pid`. Best-effort: errors are intentionally ignored
-/// (the most common error is ESRCH = the child already exited).
-#[cfg(unix)]
-fn send_sigterm(pid: u32) {
-    use nix::sys::signal::{kill, Signal};
-    use nix::unistd::Pid;
-    // `pid` came from `Child::id()` which returns a `u32` representing a
-    // POSIX `pid_t`. PIDs are positive and well below `i32::MAX` on every
-    // supported Unix kernel (Linux caps at `kernel.pid_max`, default 32768
-    // / 4 194 304; macOS at 99 998), so the bit-pattern reinterpretation
-    // via `cast_signed` is faithful — we just have to pick a signed
-    // version explicitly so clippy's `cast_possible_wrap` lint stops
-    // flagging the conversion. The workspace forbids `unsafe_code` so we
-    // go through `nix`'s safe wrapper rather than `libc::kill`.
-    let _ = kill(Pid::from_raw(pid.cast_signed()), Signal::SIGTERM);
-}
-
-#[cfg(not(unix))]
-fn send_sigterm(_pid: u32) {
-    // Windows lacks POSIX signals; the equivalent shutdown path requires
-    // `TerminateProcess`, which has different semantics (no graceful window).
-    // Until the lifecycle module is taught to switch between models, this
-    // build target relies on the child exiting via stdin EOF, and
-    // `send_sigkill` below is also a no-op.
-}
-
-/// Send `SIGKILL` to `pid`. Best-effort — ESRCH is ignored.
-#[cfg(unix)]
-fn send_sigkill(pid: u32) {
-    use nix::sys::signal::{kill, Signal};
-    use nix::unistd::Pid;
-    // Same `pid_t` reasoning as `send_sigterm` — see that function for
-    // the cast / `unsafe_code = "forbid"` rationale.
-    let _ = kill(Pid::from_raw(pid.cast_signed()), Signal::SIGKILL);
-}
-
-#[cfg(not(unix))]
-fn send_sigkill(_pid: u32) {
-    // See `send_sigterm` for the Windows scope note.
-}
-
-/// Result of [`is_sdk_compatible`]. Carried as a small enum so the caller
-/// can build the `hello_ack` reason field and the `IncompatibleSdk` error
-/// from the same source of truth.
-enum Compatibility {
-    Compatible,
-    Incompatible { reason: String },
-}
-
-impl Compatibility {
-    fn is_compatible(&self) -> bool {
-        matches!(self, Self::Compatible)
-    }
-
-    fn reason(&self) -> Option<&str> {
-        match self {
-            Self::Compatible => None,
-            Self::Incompatible { reason } => Some(reason.as_str()),
-        }
-    }
-}
-
-/// Decide whether the Bridge accepts `version` as a peer SDK.
-///
-/// The accepted set is `MAJOR.MINOR.PATCH` triples whose `MAJOR` is listed
-/// in [`ACCEPTED_SDK_MAJORS`]. Anything that fails to parse — empty string,
-/// non-numeric segments, fewer than three segments — is rejected with a
-/// reason describing the failure mode so the driver-side log makes the
-/// mismatch obvious.
-fn is_sdk_compatible(version: &str) -> Compatibility {
-    let mut parts = version.split('.');
-    let Some(major_str) = parts.next() else {
-        return Compatibility::Incompatible {
-            reason: format!("sdk_version `{version}` is empty"),
-        };
-    };
-    let Ok(major) = major_str.parse::<u32>() else {
-        return Compatibility::Incompatible {
-            reason: format!("sdk_version `{version}` has non-numeric major component"),
-        };
-    };
-    // Require the remaining `MINOR.PATCH` segments to exist, even if we don't
-    // gate on their values — the protocol contract is "MAJOR.MINOR.PATCH".
-    if parts.next().is_none() || parts.next().is_none() {
-        return Compatibility::Incompatible {
-            reason: format!("sdk_version `{version}` is not in MAJOR.MINOR.PATCH form"),
-        };
-    }
-    if !ACCEPTED_SDK_MAJORS.contains(&major) {
-        return Compatibility::Incompatible {
-            reason: format!(
-                "sdk_version `{version}` advertises major {major}, Bridge accepts {ACCEPTED_SDK_MAJORS:?}"
-            ),
-        };
-    }
-    Compatibility::Compatible
-}
-
-/// Encode and send a `hello_ack` line on `stdin`.
-fn write_hello_ack(stdin: &mut ChildStdin, ack: &HelloAckMessage<'_>) -> std::io::Result<()> {
-    serde_json::to_writer(&mut *stdin, ack)?;
-    stdin.write_all(b"\n")?;
-    stdin.flush()
-}
-
-/// Kill the child and wait for it so callers do not leak zombies on the
-/// error paths. Errors during kill / wait are intentionally swallowed: the
-/// outer `Err(SpawnError::…)` already carries the user-visible cause and
-/// the cleanup is best-effort.
-fn kill_and_reap(child: &mut Child) {
-    let _ = child.kill();
-    let _ = child.wait();
-}
-
-/// Synthesize a `serde_json::Error` for the "valid JSON but wrong tag" case
-/// so [`SpawnError::HandshakeMalformed`] can carry the diagnostic without a
-/// dedicated variant.
-fn json_type_error(message: &str) -> serde_json::Error {
-    // serde_json does not expose `Error::custom` publicly; round-trip
-    // through `serde::de::Error` via a deserializer that produces our text.
-    use serde::de::Error as _;
-    serde_json::Error::custom(message)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_should_strip_trailing_newline_when_converting_non_json_line() {
-        let event = non_json_line_to_log_event("dummy", "hello world\n");
-        assert_eq!(event.layer, "driver/dummy");
-        assert_eq!(event.device, None);
-        assert_eq!(event.message, "hello world");
-    }
-
-    #[test]
-    fn it_should_strip_trailing_crlf_when_converting_non_json_line() {
-        let event = non_json_line_to_log_event("midi", "boom\r\n");
-        assert_eq!(event.message, "boom");
-    }
-
-    #[test]
-    fn it_should_preserve_internal_whitespace_when_converting_non_json_line() {
-        let event = non_json_line_to_log_event("osc", "  spaced  message  ");
-        assert_eq!(event.message, "  spaced  message  ");
-    }
-
-    #[test]
-    fn it_should_use_driver_layer_prefix_in_log_event() {
-        let event = non_json_line_to_log_event("my-driver", "x");
-        assert_eq!(event.layer, "driver/my-driver");
-    }
 }
